wpt_babel/content-security-policy/generic/eval-typecheck-callout-order.tentative.html

<!DOCTYPE HTML>
<html>
 <head>
  <script nonce="abc" src="/resources/testharness.js">
  </script>
  <script nonce="abc" src="/resources/testharnessreport.js">
  </script>
  <title>
   Test for order of Type(evalInput) and host callout
  </title>
 </head>
 <body>
  <div id="log">
  </div>
  <script nonce="abc">
   "use strict";

test(function () {
  assert_throws_js(EvalError, function () {
    eval("0");
  }, "eval of a string should reach host callout");
}, "eval of a string should be checked by CSP");
test(function () {
  var array = ["0"];
  assert_equals(eval(array), array, "eval is identity when applied to non-strings");
}, "eval of a non-string should not be checked by CSP");
  </script>
 </body>
</html>
